Update patch set 1
Patch Set 1: Verified; Looks good to me, approved

Sorry for the rushed approval - from tomorrow I'll be traveling for a week or more. I may not have time to log again for a while, and I'm afraid this commit may miss a 7.3 release if I don't verify/approve now.

I released an unofficial build on the 1st, using the updated kernel in this commit, and users have reported that it solves overnight sleep of death issues. This never seemed to affect my particular model (GT-I5500, Europe/Ireland).

Patch-set: 1